requirements.txt
pandas>=1.5.0
numpy>=1.23.0
ccxt>=4.0.0
ta>=0.10.0
pyyaml>=6.0
pyarrow>=12.0.0